requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.6
PyYAML==6.0.2
